+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.36;
-const unsigned char AutoSDKVersionString[] = "1.36.1";
+double AutoSDKVersionNumber = 1.37;
+const unsigned char AutoSDKVersionString[] = "1.37.0";